+++ src/modules/person/PersonTest.cpp
@@ -36,7 +36,7 @@
 #include "data/macedonia/MacedonianPeopleNames.h"
 #include "data/mexico/MexicanPeopleNames.h"
 #include "data/Nationalities.h"
-#include "data/nederlands/DutchPeopleNames.h"
+#include "data/netherlands/DutchPeopleNames.h"
 #include "data/nepal/NepalesePeopleNames.h"
 #include "data/norway/NorwegianPeopleNames.h"
 #include "data/poland/PolishPeopleNames.h"
@@ -78,7 +78,7 @@ const std::map<Country, PeopleNames> countryToPeopleNamesMapping{
     {Country::Hungary, hungarianPeopleNames},     {Country::Croatia, croatianPeopleNames},
     {Country::Greece, greekPeopleNames},          {Country::Slovenia, slovenianPeopleNames},
     {Country::Austria, austrianPeopleNames},      {Country::Switzerland, swissPeopleNames},
-    {Country::Belgium, belgianPeopleNames},       {Country::Nederlands, dutchPeopleNames},
+    {Country::Belgium, belgianPeopleNames},       {Country::Netherlands, dutchPeopleNames},
     {Country::China, chinesePeopleNames},         {Country::Korea, koreanPeopleNames},
     {Country::Canada, canadianPeopleNames},       {Country::Mexico, mexicanPeopleNames},
     {Country::Argentina, argentinianPeopleNames}, {Country::Australia, australianPeopleNames},
@@ -103,7 +103,7 @@ const std::map<Country, std::string> generatedTestName{
     {Country::Hungary, "shouldGenerateHungarianName"},     {Country::Croatia, "shouldGenerateCroatianName"},
     {Country::Greece, "shouldGenerateGreekName"},          {Country::Slovenia, "shouldGenerateSlovenianName"},
     {Country::Austria, "shouldGenerateAustrianName"},      {Country::Switzerland, "shouldGenerateSwissName"},
-    {Country::Belgium, "shouldGenerateBelgianName"},       {Country::Nederlands, "shouldGenerateDutchName"},
+    {Country::Belgium, "shouldGenerateBelgianName"},       {Country::Netherlands, "shouldGenerateDutchName"},
     {Country::China, "shouldGenerateChineseName"},         {Country::Korea, "shouldGenerateKoreanName"},
     {Country::Canada, "shouldGenerateCanadianName"},       {Country::Mexico, "shouldGenerateMexicanName"},
     {Country::Argentina, "shouldGenerateArgentinianName"}, {Country::Australia, "shouldGenerateAustralianName"},
